+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
+## 2.0.0 - 14/03/2021
+* Add server side encryption to bucket objects
+* Create a user to be used when provisioning infrastructure
+
 ## 1.0.0 - 21/12/2019
 * Release to Terraform registry
 
-### 0.1.0 - 13/10/2019
+## 0.1.0 - 13/10/2019
 * Initial commit